Noiseの変更

+++ src/suan/mydns/jp/music/Wave.java
@@ -175,19 +175,19 @@ public class Wave extends PApplet
 			Volumes[3] = (VolumeR+0.5) * 1.0;
 		}
 
-		for(int i = (int) (Nokori / Frequencyss[3]); i < b.length / Frequencyss[3]; i++)
+		for(int i = (int) (Nokori / /*(int)*/Frequencyss[3]); i < b.length / /*(int)*/Frequencyss[3]; i++)
 		{
 			reg >>>= 1;
 			reg |= ((reg ^ (reg >>> (Duty == 1.0f ? 6 : 1))) & 1) << 15;
-			b[(int) (i * Frequencyss[3])] = (byte) (((reg & 1) - 0.5) * 2 * Math.max(Math.min(((byte)(Volumes[3])*8), VolumeDownUp < 0 ? 127 : MVolDUM[3] == 16 ? 127 : MVolDUM[3] * 8), VolumeDownUp >= 0 ? 0 : MVolDUM[3] == 16 ? 127 : MVolDUM[3] * 8));
-        	for(int j = 1; j < Frequencyss[3]; j++)
+			b[(int) (i * /*(int)*/Frequencyss[3])] = (byte) (((reg & 1) - 0.5) * 2 * Math.max(Math.min(((byte)(Volumes[3])*8), VolumeDownUp < 0 ? 127 : MVolDUM[3] == 16 ? 127 : MVolDUM[3] * 8), VolumeDownUp >= 0 ? 0 : MVolDUM[3] == 16 ? 127 : MVolDUM[3] * 8));
+        	for(int j = 1; j < /*(int)*/Frequencyss[3]; j++)
         	{
     			Volumes[3] = Math.max(Math.min(Volumes[3] + VolumeDownUp, 16), 0);
-        		if(i * (int)Frequencyss[3] + j < MM2.onecool) b[(int) (i * (int)Frequencyss[3] + j)] = b[(int) (i * (int)Frequencyss[3])];
+        		if(i * /*(int)*/Frequencyss[3] + j < MM2.onecool) b[(int) (i * /*(int)*/Frequencyss[3] + j)] = b[(int) (i * /*(int)*/Frequencyss[3])];
         		else
         		{
         			TempN = b[b.length-1];
-        			Nokori = i * (int)Frequencyss[3] + j - MM2.onecool;
+        			Nokori = (int)(i * /*(int)*/Frequencyss[3] + j - MM2.onecool);
         		}
         	}
 			Volumes[3] = Math.max(Math.min(Volumes[3] + VolumeDownUp, 16), 0);
